requirements.txt
streamlit>=1.35.0
msal>=1.28.0
requests>=2.31.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
pandas>=2.0.0
